Moved binary_ifstream to public interface
