+++ MagneticField/Interpolation/src/RectangularCartesianMFGrid.h
@@ -4,10 +4,13 @@
 #include "FWCore/Utilities/interface/Visibility.h"
 #include "MFGrid3D.h"
 
-class binary_ifstream;
+namespace magneticfield::interpolation {
+  class binary_ifstream;
+}
 
 class dso_internal RectangularCartesianMFGrid : public MFGrid3D {
 public:
+  using binary_ifstream = magneticfield::interpolation::binary_ifstream;
   RectangularCartesianMFGrid(binary_ifstream& istr, const GloballyPositioned<float>& vol);
 
   LocalVector uncheckedValueInTesla(const LocalPoint& p) const override;
