+++ MagneticField/Interpolation/src/RectangularCylindricalMFGrid.h
@@ -4,10 +4,14 @@
 #include "FWCore/Utilities/interface/Visibility.h"
 #include "MFGrid3D.h"
 
-class binary_ifstream;
+namespace magneticfield::interpolation {
+  class binary_ifstream;
+}
 
 class dso_internal RectangularCylindricalMFGrid : public MFGrid3D {
 public:
+  using binary_ifstream = magneticfield::interpolation::binary_ifstream;
+
   RectangularCylindricalMFGrid(binary_ifstream& istr, const GloballyPositioned<float>& vol);
 
   LocalVector uncheckedValueInTesla(const LocalPoint& p) const override;
