+++ MagneticField/Interpolation/src/TrapezoidalCartesianMFGrid.cc
@@ -4,7 +4,7 @@
  */
 
 #include "TrapezoidalCartesianMFGrid.h"
-#include "binary_ifstream.h"
+#include "MagneticField/Interpolation/interface/binary_ifstream.h"
 #include "LinearGridInterpolator3D.h"
 #include "MagneticField/VolumeGeometry/interface/MagExceptions.h"
 #include <iostream>
