+++ MagneticField/Interpolation/src/TrapezoidalCartesianMFGrid.h
@@ -14,10 +14,14 @@
 #include "Trapezoid2RectangleMappingX.h"
 #include "FWCore/Utilities/interface/Visibility.h"
 
-class binary_ifstream;
+namespace magneticfield::interpolation {
+  class binary_ifstream;
+}
 
 class dso_internal TrapezoidalCartesianMFGrid : public MFGrid3D {
 public:
+  using binary_ifstream = magneticfield::interpolation::binary_ifstream;
+
   TrapezoidalCartesianMFGrid(binary_ifstream& istr, const GloballyPositioned<float>& vol);
 
   LocalVector uncheckedValueInTesla(const LocalPoint& p) const override;
